Switch to LWP::UserAgent instead of HTTP::Tiny
HTTP::Tiny requires newer versions of SSL modules to work properly, but
on our production CentOS box, LWP::UserAgent is used in the majority of
modules and it needs older versions of the same module.

We could install a the new versions of the module via local::lib but
it's a pain to remember that this is inconsistent with the rest of the
modules.